pyproject.toml
[project]
name = "reverse-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "demucs>=4.0.1",
    "faster-whisper>=1.2.1",
    "google-genai>=1.70.0",
    "opentimelineio>=0.18.1",
    "otio-fcp-adapter>=1.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "scenedetect[opencv]>=0.6.7.1",
    "torchcodec>=0.11.0",
    "yt-dlp>=2026.3.17",
    "numpy>=2.0",
    "scikit-learn>=1.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
]
